requirements.txt
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-compute==1.19.2
google-cloud-vpc-access==1.10.5
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
idna==3.8
prettytable==3.11.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
wcwidth==0.2.13
google-api-python-client